Dia/diabetes_repo/requirements.txt
numpy
pandas
scikit-learn
torch
tqdm
pyyaml
joblib
